DBへ指示を送るクラスの名称及び内部コードの構成を変更
